Fix 4 failing seeds: Illusion position, Fillet Away EmergencyExit, PRNG u32 support, and Polkadot Bow
Fixes seeds 29000, 40192, 47174, 87989

1. Illusion ability (seed 29000):
   - Fixed to use pokemon_index (party index 0-5) instead of pokemon.position
   - In JavaScript, pokemon.position refers to party position, not active slot

2. Fillet Away move (seed 47174):
   - Removed incorrect EmergencyExit check after directDamage
   - In JavaScript, directDamage does NOT trigger EmergencyExit

3. PRNG seed handling (seed 87989):
   - Changed PRNGSeed::Gen5 from [u16; 4] to [u32; 4] for seeds > 65535
   - Updated all examples to pass full u32 seed values

4. Polkadot Bow item (seed 87989):
   - Fixed to match JavaScript's floating-point precision behavior
   - Returns Number for exact integers (10*1.1=11), Float for non-integers
   - Ensures Rivalry modifier applies correctly based on result type

All 500 regression test seeds pass.

+++ examples/generate_test_teams_rust.rs
@@ -41,7 +41,8 @@ struct StatsExport {
 }
 
 fn main() {
-    let seed_num: u16 = env::args()
+    // Parse seed as u32 to handle large values (matching JavaScript behavior)
+    let seed_num: u32 = env::args()
         .nth(1)
         .and_then(|s| s.parse().ok())
         .unwrap_or(1);
@@ -51,7 +52,7 @@ fn main() {
     // Load dex
     let dex = Dex::load_default().expect("Failed to load dex");
 
-    // Create PRNG with seed
+    // Create PRNG with seed (pass full u32 value, masking happens in PRNG operations)
     let mut prng = PRNG::new(Some(PRNGSeed::Gen5([0, 0, 0, seed_num])));
 
     // Generate teams
